#39 グラデーションの関数ファイルを追加
