+++ kyutechapp2017/BulletinBoard.storyboard
@@ -132,7 +132,7 @@
                                 </constraints>
                             </stackView>
                         </subviews>
-                        <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="calibratedRGB"/>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
                         <constraints>
                             <constraint firstItem="MJv-cS-Be3" firstAttribute="top" secondItem="MyK-Sg-H3J" secondAttribute="bottom" constant="8" id="79j-bl-2XI"/>
                             <constraint firstItem="Iyg-cp-65D" firstAttribute="top" secondItem="MJv-cS-Be3" secondAttribute="bottom" constant="8" id="Gc6-LK-o0U"/>
@@ -142,13 +142,14 @@
                     </view>
                     <navigationItem key="navigationItem" title="Kyutech" id="oKK-W4-VZk"/>
                     <connections>
-                        <outlet property="StudyAbroadButton" destination="U9t-fd-tWQ" id="XvE-hK-gNx"/>
+                        <outlet property="BackgroundView" destination="XQJ-MJ-PvE" id="wzs-p9-5kI"/>
                         <outlet property="classesButton" destination="kxp-Qd-AbJ" id="vcR-QW-IFD"/>
                         <outlet property="departmentButton" destination="4xs-L2-Z64" id="WaE-Ij-GEj"/>
                         <outlet property="intensiveButton" destination="Web-0C-FB8" id="nWT-rq-lrv"/>
                         <outlet property="newsButton" destination="Py0-ss-FIw" id="Xq3-GT-mE0"/>
                         <outlet property="procedureButton" destination="8IK-g7-wUF" id="6HT-63-zkT"/>
                         <outlet property="scholarshipButton" destination="yNa-Ff-f5A" id="F3N-gb-I4G"/>
+                        <outlet property="studyAbroadButton" destination="U9t-fd-tWQ" id="9Ug-6L-NmQ"/>
                         <outlet property="summonButton" destination="ifv-4l-4bq" id="GVR-4p-B4r"/>
                     </connections>
                 </viewController>
